refactor:useAuthToken 분리
