+++ my_blog/src/App.tsx
@@ -1,7 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { useEffect } from 'react';
-import axiosInstance from './api/axiosInstance';
 import { ToastProvider } from './context/ToastContext';
+import { useAuthToken } from './hooks/useAuthToken';
 import {
   TestForBlank,
   TestForButton,
@@ -34,31 +33,8 @@ import OAuthSignUp from './pages/auth/OAthSignUp';
 import KakaoRedirect from './pages/auth/KakaoRedirect';
 
 function App() {
-  useEffect(() => {
-    const token = localStorage.getItem('accessToken');
-    const refreshToken = localStorage.getItem('refreshToken');
+  useAuthToken();
 
-    if (token) {
-      axiosInstance.defaults.headers.common.Authorization = `Bearer ${token}`;
-    } else if (refreshToken) {
-      const refreshAccessToken = async () => {
-        try {
-          const res = await axiosInstance.post('/auth/reissue', {
-            refreshToken,
-          });
-          const newToken = res.data.data.accessToken;
-          localStorage.setItem('accessToken', newToken);
-          axiosInstance.defaults.headers.common.Authorization = `Bearer ${newToken}`;
-        } catch (err) {
-          console.error('토큰 갱신 실패:', err);
-          localStorage.removeItem('accessToken');
-          localStorage.removeItem('refreshToken');
-          window.location.href = '/';
-        }
-      };
-      refreshAccessToken();
-    }
-  }, []);
   return (
     <ToastProvider>
       <BrowserRouter>
